refactor(core): replacing the main bootstrap to files and folders
