+++ backend/individual-assignment/src/models/PaymentMethod.ts
@@ -1,0 +1,1 @@
+export default interface PaymentMethod { }
