+++ backend/individual-assignment/src/services/address.ts
@@ -1,0 +1,1 @@
+export default class Address { }
