+++ backend/individual-assignment/src/services/credit-card.ts
@@ -1,0 +1,5 @@
+import PaymentMethod from "../models/PaymentMethod";
+
+export default class CreditCard implements PaymentMethod {
+  constructor(public number: string) { }
+}
