+++ backend/individual-assignment/src/services/customer.ts
@@ -1,0 +1,1 @@
+export default class Customer { }
