+++ backend/individual-assignment/src/services/order-item.ts
@@ -1,0 +1,9 @@
+import Product from "./product";
+
+export default class OrderItem {
+  constructor(public product: Product, public quantity: number) { }
+
+  get total(): number {
+    return this.product.price * this.quantity;
+  }
+}
